@@ -2,29 +2,25 @@ import re
 import socket
 import sys
 from contextlib import contextmanager
+from typing import Any, Callable, Iterator, List, Optional, Tuple, Union
+from urllib.parse import urlparse
 
 import attr
-
-try:
-    from urllib.parse import urlparse
-except ImportError:
-    from urlparse import urlparse
-
 
 try:
     import pycurl
 
     @attr.s(slots=True)
-    class Curl(object):
+    class Curl:
         """Proxy to real pycurl.Curl.
 
         If `perform` is called then it will raise an error if network is disabled via `disable`
         """
 
-        handle = attr.ib(factory=pycurl.Curl)
-        url = None
+        handle = attr.ib(factory=pycurl.Curl, type=pycurl.Curl)
+        url = None  # type: Optional[str]
 
-        def __getattribute__(self, item):
+        def __getattribute__(self, item: str) -> Any:
             handle = object.__getattribute__(self, "handle")
             if _disable_pycurl and item == "perform":
                 host = urlparse(self.url).hostname
@@ -37,21 +33,21 @@ try:
                 return object.__getattribute__(self, "setopt")
             return getattr(handle, item)
 
-        def __setattr__(self, key, value):
+        def __setattr__(self, key: str, value: Any) -> None:
             if key == "handle":
                 object.__setattr__(self, key, value)
             else:
                 setattr(self.handle, key, value)
 
-        def setopt(self, option, value):
+        def setopt(self, option: int, value: Any) -> None:
             if option == pycurl.URL:
                 self.url = value
             self.handle.setopt(option, value)
 
 
 except ImportError:
-    pycurl = None
-    Curl = None
+    pycurl = None  # type: ignore
+    Curl = None  # type: ignore
 
 # `socket.socket` is not patched, because it could be needed for live servers (e.g. pytest-httpbin)
 # But methods that could connect to remote are patched to prevent network access
@@ -64,67 +60,67 @@ _allowed_hosts = None
 
 
 @attr.s(slots=True, hash=True)
-class PyCurlWrapper(object):
+class PyCurlWrapper:
     """Imitate pycurl module."""
 
-    def __getattribute__(self, item):
+    def __getattribute__(self, item: str) -> Any:
         if item == "Curl":
             return Curl
         return getattr(pycurl, item)
 
 
-def check_pycurl_installed(func):
+def check_pycurl_installed(func: Callable) -> Callable:
     """No-op if pycurl is not installed."""
 
-    def inner(*args, **kwargs):  # pylint: disable=inconsistent-return-statements
+    def inner(*args: Any, **kwargs: Any) -> Any:  # pylint: disable=inconsistent-return-statements
         if pycurl is None:
-            return
+            return  # type: ignore
         return func(*args, **kwargs)
 
     return inner
 
 
 @check_pycurl_installed
-def install_pycurl_wrapper():
-    sys.modules["pycurl"] = PyCurlWrapper()
+def install_pycurl_wrapper() -> None:
+    sys.modules["pycurl"] = PyCurlWrapper()  # type: ignore
 
 
 @check_pycurl_installed
-def uninstall_pycurl_wrapper():
+def uninstall_pycurl_wrapper() -> None:
     sys.modules["pycurl"] = pycurl
 
 
-def block_pycurl(allowed_hosts=None):
+def block_pycurl(allowed_hosts: Optional[List[str]] = None) -> None:
     global _disable_pycurl  # pylint: disable=global-statement
     global _allowed_hosts  # pylint: disable=global-statement
     _disable_pycurl = True
     _allowed_hosts = allowed_hosts
 
 
-def unblock_pycurl():
+def unblock_pycurl() -> None:
     global _disable_pycurl  # pylint: disable=global-statement
     global _allowed_hosts  # pylint: disable=global-statement
     _disable_pycurl = False
     _allowed_hosts = None
 
 
-def block_socket(allowed_hosts=None):
-    socket.socket.connect = make_network_guard(_original_connect, allowed_hosts=allowed_hosts)
-    socket.socket.connect_ex = make_network_guard(_original_connect_ex, allowed_hosts=allowed_hosts)
+def block_socket(allowed_hosts: Optional[List[str]] = None) -> None:
+    socket.socket.connect = make_network_guard(_original_connect, allowed_hosts=allowed_hosts)  # type: ignore
+    socket.socket.connect_ex = make_network_guard(_original_connect_ex, allowed_hosts=allowed_hosts)  # type: ignore
 
 
-def unblock_socket():
-    socket.socket.connect = _original_connect
-    socket.socket.connect_ex = _original_connect_ex
+def unblock_socket() -> None:
+    socket.socket.connect = _original_connect  # type: ignore
+    socket.socket.connect_ex = _original_connect_ex  # type: ignore
 
 
-def make_network_guard(original_func, allowed_hosts=None):
-    def network_guard(self, address, *args, **kwargs):
-        host = ""
+def make_network_guard(original_func: Callable, allowed_hosts: Optional[List[str]] = None) -> Callable:
+    def network_guard(self: Any, address: Union[Tuple, str, bytes], *args: Any, **kwargs: Any) -> Any:
+        host = ""  # type: Union[str, bytes, bytearray]
         if self.family in (socket.AF_INET, socket.AF_INET6):
-            host = address[0]
+            host = address[0]  # type: ignore
         elif self.family == socket.AF_UNIX:
-            host = address
+            host = address  # type: ignore
         if is_host_in_allowed_hosts(host, allowed_hosts):
             return original_func(self, address, *args, **kwargs)
         raise RuntimeError("Network is disabled")
@@ -132,19 +128,19 @@ def make_network_guard(original_func, allowed_hosts=None):
     return network_guard
 
 
-def block(allowed_hosts=None):
+def block(allowed_hosts: Optional[List[str]] = None) -> None:
     block_socket(allowed_hosts=allowed_hosts)
     # NOTE: Applying socket blocking makes curl hangs - it should be carefully patched
     block_pycurl(allowed_hosts=allowed_hosts)
 
 
-def unblock():
+def unblock() -> None:
     unblock_pycurl()
     unblock_socket()
 
 
 @contextmanager
-def blocking_context(allowed_hosts=None):
+def blocking_context(allowed_hosts: Optional[List[str]] = None) -> Iterator[None]:
     """Block connections via socket and pycurl.
 
     Note:
@@ -160,9 +156,15 @@ def blocking_context(allowed_hosts=None):
         unblock()
 
 
-def is_host_in_allowed_hosts(host, allowed_hosts):
+def to_string(value: Union[str, bytes, bytearray]) -> str:
+    if isinstance(value, (bytes, bytearray)):
+        return value.decode()
+    return value
+
+
+def is_host_in_allowed_hosts(host: Union[str, bytes, bytearray], allowed_hosts: Optional[List[str]]) -> bool:
     """Match provided host to a list of host regexps."""
     if allowed_hosts is not None:
         combined = "(" + ")|(".join(allowed_hosts) + ")"
-        return re.match(combined, host)
+        return bool(re.match(combined, to_string(host)))
     return False
