@@ -1,7 +1,8 @@
 from itertools import chain
+from typing import Any, Iterable, Iterator
 
 
-def unique(sequence):
+def unique(sequence: Iterable) -> Iterator:
     seen = set()
     for item in sequence:
         if item not in seen:
@@ -9,5 +10,5 @@ def unique(sequence):
             yield item
 
 
-def unpack(*args):
+def unpack(*args: Any) -> Iterable:
     return chain.from_iterable(args)
